requirements.txt
selenium==4.18.1
pytest==8.1.1
webdriver-manager==4.0.1
allure-pytest==2.13.2
requests